Consolidates broken-out utils files in assets/js
