+++ assets/js/utility/vrms-events.mjs
@@ -1,8 +1,6 @@
 ---
 ---
 
-import { filterTestEvents, sortEventsByDate } from "./vrms-events-utils.mjs";
-
 {% assign vrmsData = site.data.external.vrms_data %}
 const vrmsData = JSON.parse(decodeURIComponent("{{ vrmsData | jsonify | uri_escape }}"));
 
@@ -18,6 +16,28 @@ export const vrmsDataFetch = (currentPage, appendMeetingTimes) => {
   else if (currentPage == "project") {
     appendMeetingTimes(sortedEvents);
   }
+}
+
+/**
+ * Sorts an array of events by their "startTime" variable, from oldest to newest
+ */
+export function sortEventsByDate(events) {
+     return events.sort((event1, event2) => {
+        // startTime includes date information
+        const date1 = new Date(event1.startTime);
+        const date2 = new Date(event2.startTime);
+        return date1 - date2;
+     });
+}
+
+/**
+ * Filters an array of events to remove instances of events with names containing
+ * "test" or "testing, case insensitive
+ */
+export function filterTestEvents(events) {
+    return events.filter(event => {
+        return !(/\btest(ing)?\b/i.test(event.name))
+    });
 }
 
 /**
